requirements.txt
anytree
appdirs
aria2p
asyncio
attrdict
beautifulsoup4
bencoding
cfscrape
configparser
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
pybase64
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
telegraph
tenacity
tgCrypto
urllib3
yt-dlp
